# AngularGui/src/main/java/app/stock-view-app/src/app/model/SearchCriteria.ts
/* tslint:disable:max-line-length */
/**
 * App with angular GIU and springboot backend and swagger description for REST
 * 1.0
 * Swagger stock app
 * localhost:9091
 */

export interface SearchCriteria {
  /** format: float */
  maxPrice?: number;
  /** format: float */
  minPrice?: number;
}