# SpringBoot/src/main/java/app/model/SearchCriteria.java
package app.model;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class SearchCriteria {

    private Float minPrice;
    private Float maxPrice;
}